# services/gateway/swagger.yaml
openapi: 3.0.0
info:
  title: AI Code Reviewer API
  description: |
    API complète pour l'analyse automatique de code source via Intelligence Artificielle.
    
    **Architecture**: Microservices avec API Gateway
    
    **Services**:
    - Auth Service (Port 3001) - Authentification JWT
    - Code Service (Port 3002) - Gestion des analyses
    - AI Service (Port 3003) - Analyse OpenAI
    - Payment Service (Port 3004) - Paiements Stripe
    
    **Authentication**: Utiliser le header `Authorization: Bearer <token>`
  version: 1.0.0
  contact:
    name: Support AI Code Reviewer
    email: support@aicodereviewer.com
servers:
  - url: http://localhost:3000/api
    description: Développement local
  - url: https://api.aicodereviewer.com/api
    description: Production
...
paths:
...
  /payment/create-checkout:
    post:
      summary: Créer une session de paiement Stripe
      tags: [Paiement]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan]
              properties:
                plan:
                  type: string
                  enum: [monthly, annual, lifetime]
                  example: "monthly"
                successUrl:
                  type: string
                  format: uri
                cancelUrl:
                  type: string
                  format: uri
      responses:
        200:
          description: Session créée
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      sessionId:
                        type: string
                      checkoutUrl:
                        type: string
                        format: uri
...
  /payment/subscription:
    get:
      summary: Récupérer l'abonnement actuel
      tags: [Paiement]
      security:
        - BearerAuth: []
      responses:
        200:
          description: Informations de l'abonnement
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      hasSubscription:
                        type: boolean
                      plan:
                        type: string
                      status:
                        type: string
                      isActive:
                        type: boolean
                      currentPeriodEnd:
                        type: string
                        format: date-time
  /payment/cancel-subscription:
    delete:
      summary: Annuler l'abonnement
      tags: [Paiement]
      security:
        - BearerAuth: []
      responses:
        200:
          description: Abonnement annulé
  /payment/invoices:
    get:
      summary: Historique des factures
      tags: [Paiement]
      security:
        - BearerAuth: []
      responses:
        200:
          description: Liste des factures
  /payment/portal:
    post:
      summary: Accéder au portail client Stripe
      tags: [Paiement]
      security:
        - BearerAuth: []
      responses:
        200:
          description: URL du portail

  # ==================== HEALTH ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu lors de la connexion
...